Review feedback: move `signer.rs` into `internal.rs`, and rename `_remove_key()` to `remove_key()` as it IS used by the RTA code. Also make key removal code more consistent across signer impls and always remove the key from the `SignerMapper` too.
