+++ src/commons/crypto/signing/signers/kmip/mod.rs
@@ -15,6 +15,5 @@
 //! [PyKMIP]: https://github.com/OpenKMIP/PyKMIP
 pub mod connpool;
 pub mod internal;
-pub mod signer;
 
 pub use internal::KmipSigner;
